+++ src/mako/pixel/image/operations/gd/Resize.php
@@ -33,7 +33,7 @@ class Resize implements OperationInterface
 	 */
 	public function __construct(
 		protected int $width,
-		protected ?int $height = null,
+		protected int $height,
 		protected AspectRatio $aspectRatio = AspectRatio::Auto
 	) {
 	}