+++ tests/unit/pixel/image/GdTest.php
@@ -9,8 +9,6 @@ namespace mako\tests\unit\pixel\image;
 
 use mako\pixel\image\exceptions\ImageException;
 use mako\pixel\image\Gd;
-use mako\pixel\image\operations\gd\Bitonal;
-use mako\pixel\image\operations\gd\Negate;
 use mako\tests\TestCase;
 use PHPUnit\Framework\Attributes\Group;
 
@@ -39,72 +37,6 @@ class GdTest extends TestCase
 		$this->assertSame(300, $image->getHeight());
 
 		$this->assertSame(['width' => 300, 'height' => 300], $image->getDimensions());
-	}
-
-	/**
-	 * Note that the colors extracted by GD aren't 100% accurate.
-	 */
-	public function testGetTopColors(): void
-	{
-		$image = new Gd(__DIR__ . '/fixtures/001.png');
-
-		$colors = $image->getTopColors();
-
-		$this->assertCount(3, $colors);
-
-		$this->assertSame('#0070C0', $colors[0]->toHexString());
-		$this->assertSame('#B01000', $colors[1]->toHexString());
-		$this->assertSame('#007000', $colors[2]->toHexString());
-	}
-
-	/**
-	 * Note that the colors extracted by GD aren't 100% accurate.
-	 */
-	public function testGetTopColorsWithLimit(): void
-	{
-		$image = new Gd(__DIR__ . '/fixtures/001.png');
-
-		$colors = $image->getTopColors(2);
-
-		$this->assertCount(2, $colors);
-
-		$this->assertSame('#0070C0', $colors[0]->toHexString());
-		$this->assertSame('#B01000', $colors[1]->toHexString());
-	}
-
-	/**
-	 *
-	 */
-	public function testBitonal(): void
-	{
-		$image = new Gd(__DIR__ . '/fixtures/002.jpg');
-
-		$image->apply(new Bitonal);
-
-		$colors = $image->getTopColors();
-
-		$this->assertSame(2, count($colors));
-
-		$this->assertSame('#FFFFFF', $colors[0]->toHexString());
-		$this->assertSame('#000000', $colors[1]->toHexString());
-	}
-
-	/**
-	 *
-	 */
-	public function testNegate(): void
-	{
-		$image = new Gd(__DIR__ . '/fixtures/002.jpg');
-
-		$image->apply(new Bitonal);
-		$image->apply(new Negate);
-
-		$colors = $image->getTopColors();
-
-		$this->assertSame(2, count($colors));
-
-		$this->assertSame('#000000', $colors[0]->toHexString());
-		$this->assertSame('#FFFFFF', $colors[1]->toHexString());
 	}
 
 	/**