+++ tests/unit/pixel/image/operations/imagemagick/ResizeTest.php
@@ -1,0 +1,80 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\tests\unit\pixel\image\operations\imagemagick;
+
+use mako\pixel\image\ImageMagick;
+use mako\pixel\image\operations\AspectRatio;
+use mako\pixel\image\operations\imagemagick\Resize;
+use mako\tests\TestCase;
+use PHPUnit\Framework\Attributes\Group;
+
+#[Group('unit')]
+class ResizeTest extends TestCase
+{
+	/**
+	 *
+	 */
+	public function setUp(): void
+	{
+		if (!extension_loaded('imagick')) {
+			$this->markTestSkipped('The "imagick" extension is not enabled.');
+		}
+	}
+
+	/**
+	 *
+	 */
+	public function testResize(): void
+	{
+		$image = new ImageMagick(__DIR__ . '/../../fixtures/001.png');
+
+		$image->apply(new Resize(50, 50));
+
+		$this->assertSame(50, $image->getHeight());
+		$this->assertSame(50, $image->getWidth());
+	}
+
+	/**
+	 *
+	 */
+	public function testResizeAspectRatioHeight(): void
+	{
+		$image = new ImageMagick(__DIR__ . '/../../fixtures/001.png');
+
+		$image->apply(new Resize(100, 50, AspectRatio::Height));
+
+		$this->assertSame(50, $image->getHeight());
+		$this->assertSame(50, $image->getWidth());
+	}
+
+	/**
+	 *
+	 */
+	public function testResizeAspectRatioWidth(): void
+	{
+		$image = new ImageMagick(__DIR__ . '/../../fixtures/001.png');
+
+		$image->apply(new Resize(100, 50, AspectRatio::Width));
+
+		$this->assertSame(100, $image->getHeight());
+		$this->assertSame(100, $image->getWidth());
+	}
+
+	/**
+	 *
+	 */
+	public function testResizeAspectRatioIgnore(): void
+	{
+		$image = new ImageMagick(__DIR__ . '/../../fixtures/001.png');
+
+		$image->apply(new Resize(100, 50, AspectRatio::Ignore));
+
+		$this->assertSame(50, $image->getHeight());
+		$this->assertSame(100, $image->getWidth());
+	}
+}